+++ relayer/src/rpc_conf.rs
@@ -3,6 +3,7 @@
  */
 
 use std::fmt::Debug;
+use color_eyre::Report;
 
 #[derive(Eq, Hash, Clone, Debug, PartialEq)]
 pub struct ApiKey(pub near_jsonrpc_client::auth::ApiKey);
@@ -138,4 +139,59 @@ impl NetworkConfig {
         };
         json_rpc_client
     }
+}
+
+pub fn rpc_transaction_error(
+    err: near_jsonrpc_client::errors::JsonRpcError<
+        near_jsonrpc_client::methods::broadcast_tx_commit::RpcTransactionError,
+    >,
+) -> Result<(), Report> {
+    match &err {
+        near_jsonrpc_client::errors::JsonRpcError::TransportError(_rpc_transport_error) => {
+            println!("Transport error transaction.\nPlease wait. The next try to send this transaction is happening right now ...");
+        }
+        near_jsonrpc_client::errors::JsonRpcError::ServerError(rpc_server_error) => match rpc_server_error {
+            near_jsonrpc_client::errors::JsonRpcServerError::HandlerError(rpc_transaction_error) => match rpc_transaction_error {
+                near_jsonrpc_client::methods::broadcast_tx_commit::RpcTransactionError::TimeoutError => {
+                    println!("Timeout error transaction.\nPlease wait. The next try to send this transaction is happening right now ...");
+                }
+                near_jsonrpc_client::methods::broadcast_tx_commit::RpcTransactionError::InvalidTransaction { context } => {
+                    return Err(color_eyre::eyre::eyre!("Invalid Transaction Error. Context: {}", context.to_string()));
+                }
+                near_jsonrpc_client::methods::broadcast_tx_commit::RpcTransactionError::DoesNotTrackShard => {
+                    return Err(color_eyre::eyre::eyre!("RPC Server Error"));
+                }
+                near_jsonrpc_client::methods::broadcast_tx_commit::RpcTransactionError::RequestRouted{transaction_hash} => {
+                    return Err(color_eyre::eyre::eyre!("RPC Server Error for transaction with hash {}\n", transaction_hash));
+                }
+                near_jsonrpc_client::methods::broadcast_tx_commit::RpcTransactionError::UnknownTransaction{requested_transaction_hash} => {
+                    return Err(color_eyre::eyre::eyre!("RPC Server Error for transaction with hash {}\n", requested_transaction_hash));
+                }
+                near_jsonrpc_client::methods::broadcast_tx_commit::RpcTransactionError::InternalError{debug_info} => {
+                    return Err(color_eyre::eyre::eyre!("RPC Server Error: {}", debug_info));
+                }
+            }
+            near_jsonrpc_client::errors::JsonRpcServerError::RequestValidationError(rpc_request_validation_error) => {
+                return Err(color_eyre::eyre::eyre!("Incompatible request with the server: {:#?}",  rpc_request_validation_error));
+            }
+            near_jsonrpc_client::errors::JsonRpcServerError::InternalError{ info } => {
+                println!("Internal server error: {}.\nPlease wait. The next try to send this transaction is happening right now ...", info.clone().unwrap_or_default());
+            }
+            near_jsonrpc_client::errors::JsonRpcServerError::NonContextualError(rpc_error) => {
+                return Err(color_eyre::eyre::eyre!("Unexpected response: {}", rpc_error));
+            }
+            near_jsonrpc_client::errors::JsonRpcServerError::ResponseStatusError(json_rpc_server_response_status_error) => match json_rpc_server_response_status_error {
+                near_jsonrpc_client::errors::JsonRpcServerResponseStatusError::Unauthorized => {
+                    return Err(color_eyre::eyre::eyre!("JSON RPC server requires authentication. Please, authenticate near CLI with the JSON RPC server you use."));
+                }
+                near_jsonrpc_client::errors::JsonRpcServerResponseStatusError::TooManyRequests => {
+                    println!("JSON RPC server is currently busy.\nPlease wait. The next try to send this transaction is happening right now ...");
+                }
+                near_jsonrpc_client::errors::JsonRpcServerResponseStatusError::Unexpected{status} => {
+                    return Err(color_eyre::eyre::eyre!("JSON RPC server responded with an unexpected status code: {}", status));
+                }
+            }
+        }
+    }
+    Ok(())
 }
